fix: bet placing tests pass
