+++ contracts/MarketGreedy.sol
@@ -6,17 +6,11 @@ import "./Market.sol";
 //Import Hardhat console.log
 import "hardhat/console.sol";
 
-contract MarketGreedy is Market {
+abstract contract MarketGreedy is Market {
     // MarketID => amount of cover taken for this market
 	mapping(bytes16 => uint256) private _marketCover;
+	uint256 internal _totalCover;
 
-    constructor(
-        IVault vault,
-        uint8 fee,
-        address oracle
-    ) Market(vault, fee, oracle) {
-    }
-    
     function getMarketCover(bytes16 marketId) external view returns (uint256) {
 		return _marketCover[marketId];
 	}
@@ -32,42 +26,80 @@ contract MarketGreedy is Market {
 		if (result == true) {
 			// Send the payout to the bet owner
 			IERC20(underlying).transfer(_bets[index].owner, _bets[index].payout);
-            _totalCover -= _bets[index].payout - _bets[index].amount;
+			// Deduct from market cover
+            
 		} else {
 			// Send the bet amount to the vault
 			IERC20(underlying).transfer(address(_vault), _bets[index].amount);
+			uint256 cover = _bets[index].payout - _bets[index].amount;
+			if (_marketCover[_bets[index].marketId] >= cover) {
+				_marketCover[_bets[index].marketId] -= cover;
+				_totalExposure -= cover;
+			} else {
+				_totalExposure -= _marketCover[_bets[index].marketId];
+				_marketCover[_bets[index].marketId] = 0;
+			}		
 		}
         
 	}
 
-	// Overriden to make this "greedy" - it will hold on to the cover amounts for future bets
+	// Overridden to make this "greedy" - it will hold on to the cover amounts for future bets
     // If the payout for this proposition will be greater than the amount of cover set aside for the market
+	// Return the new exposure amount
 	function _obtainCover(bytes16 marketId, bytes16 propositionId, uint256 wager, uint256 payout) internal override returns (uint256) {
-        if (_potentialPayout[propositionId] > _marketCover[marketId]) {
+        console.log("_obtainCover() start: _payout: %s", payout);
+		uint256 result = 0;
+		console.log("_potentialPayout[propositionId]: %s", _potentialPayout[propositionId]);
+		console.log("_marketCover[marketId]: %s", _marketCover[marketId]);
+		if (_potentialPayout[propositionId] > _marketCover[marketId]) {
 			// Get any additional cover we need for this market
+
 			uint256 amountRequired =_potentialPayout[propositionId] - _marketCover[marketId];
+			//console.log("amountRequired: %s", amountRequired);
+
+			//assert(_totalCover >= _totalExposure);
+			//console.log("_totalCover: %s", _totalCover);
+			//console.log("_totalExposure: %s", _totalExposure);
             uint256 internallyAvailableCover = _totalCover - _totalExposure;
+			//console.log("internallyAvailableCover: %s", internallyAvailableCover);
+
             uint256 internalCoverToUse = Math.min(amountRequired, internallyAvailableCover);
+			//console.log("internalCoverToUse: %s", internalCoverToUse);
                     
             if (internalCoverToUse < amountRequired) {
                 // We need to get more cover from the Vault 
-                uint256 amountToTransfer = amountRequired - internalCoverToUse;           
+				//console.log("Getting more cover from the Vault");
+                uint256 amountToTransfer = amountRequired - internalCoverToUse;   
+				console.log("amountToTransfer: %s", amountToTransfer);        
                 IERC20(_vault.asset()).transferFrom(
                     address(_vault),
                     _self,
                     amountToTransfer
                 ); 
-                _totalCover += amountToTransfer;  
-                
-            } 
-            _marketCover[marketId] += amountRequired;           
+				//console.log("Cover obtained from the Vault");
+                result = amountToTransfer;
+				_totalCover += result;
+            } else {
+				result = amountRequired;
+			}
+            _marketCover[marketId] += amountRequired;   
+			console.log("_marketCover[marketId] now: %s", _marketCover[marketId]);    
 		}
+		
+		//console.log("_totalCover now: %s", _totalCover);
+		//console.log("_obtainCover() end");
+		return result;
+	}
+
+	function getTotalCover() external view returns (uint256) {
+		return _totalCover;
 	}
 
 	// Allow the Vault to reclaim any cover it has provided, provided it is not currently covering any bets
 	function reclaimCover() external onlyOwner {
-		IERC20(_vault.asset()).transfer(address(_vault), _totalCover - _totalExposure);
-		_totalCover = _totalExposure;
+		if (_totalCover > _totalExposure) {
+			IERC20(_vault.asset()).transfer(address(_vault), _totalCover - _totalExposure);
+			_totalCover = _totalExposure;		
+		}
 	}
-
 }
