+++ package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@chainlink/contracts": "^0.5.1",
     "@ethersproject/providers": "^5.0.0",
-    "@openzeppelin/contracts": "^4.8.0",
+    "@openzeppelin/contracts": "latest",
     "@typechain/ethers-v5": "^8.0.5",
     "@typechain/hardhat": "^3.1.0",
     "chai": "^4.3.4",
