+++ tests/utils.ts
@@ -1,5 +1,12 @@
-﻿import { ethers } from "ethers";
+﻿import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
+import { BigNumber, BigNumberish, ethers } from "ethers";
 import { concat, hexlify, toUtf8Bytes } from "ethers/lib/utils";
+
+type Signature = {
+	v: BigNumberish;
+	r: string;
+	s: string;
+};
 
 export const getEventData = (
 	eventName: string,
@@ -49,3 +56,105 @@ export function makeMarketId(date: Date, location: string, raceNumber: string) {
 export function makePropositionId(marketId: string, prediction: number) {
 	return `${marketId}W${prediction.toString().padStart(2, "0")}`;
 }
+
+export const signBackMessage = async (
+	nonce: string,
+	marketId: string,
+	propositionId: string,
+	odds: BigNumber,
+	close: number,
+	end: number,
+	signer: SignerWithAddress
+): Promise<Signature> => {
+	const message = ethers.utils.solidityKeccak256(
+		[
+			"bytes16", // nonce
+			"bytes16", // propositionId
+			"bytes16", // marketId
+			"uint256", // odds
+			"uint256", // close
+			"uint256" // end
+		],
+		[
+			formatBytes16String(nonce),
+			formatBytes16String(propositionId),
+			formatBytes16String(marketId),
+			odds,
+			close,
+			end
+		]
+	);
+	return await signMessage(message, signer);
+};
+
+export const signMessage = async (
+	message: string,
+	signer: SignerWithAddress
+): Promise<Signature> => {
+	const sig = await signer.signMessage(ethers.utils.arrayify(message));
+	const { v, r, s } = ethers.utils.splitSignature(sig);
+	return { v, r, s };
+};
+
+export const signSetResultMessage = async (
+	marketId: string,
+	propositionId: string,
+	signer: SignerWithAddress
+): Promise<Signature> => {
+	const settleMessage = makeSetResultMessage(marketId, propositionId);
+	return await signMessage(settleMessage, signer);
+};
+
+const makeSetResultMessage = (
+	marketId: string,
+	propositionId: string
+): string => {
+	const b16MarketId = formatBytes16String(marketId);
+	const b16PropositionId = formatBytes16String(propositionId);
+	const message = ethers.utils.solidityKeccak256(
+		["bytes16", "bytes16"],
+		[b16MarketId, b16PropositionId]
+	);
+	return message;
+};
+
+const signMessageAsString = async (
+	message: string,
+	signer: SignerWithAddress
+) => {
+	const sig = await signer.signMessage(ethers.utils.arrayify(message));
+	return sig;
+};
+
+const signBackMessageWithRisk = async (
+	nonce: string,
+	marketId: string,
+	propositionId: string,
+	odds: BigNumber,
+	close: number,
+	end: number,
+	risk: number,
+	signer: SignerWithAddress
+): Promise<Signature> => {
+	const message = ethers.utils.solidityKeccak256(
+		[
+			"bytes16", // nonce
+			"bytes16", // propositionId
+			"bytes16", // marketId
+			"uint256", // odds
+			"uint256", // close
+			"uint256", // end
+			"uint256" // risk
+		],
+		[
+			formatBytes16String(nonce),
+			formatBytes16String(propositionId),
+			formatBytes16String(marketId),
+			odds,
+			close,
+			end,
+			risk
+		]
+	);
+	return await signMessage(message, signer);
+};
